pyproject.toml
[project]
name = "technitium-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "flake8>=7.3.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "structlog>=25.4.0",
]
